feat(test): add ecdsaSign test, deprecate signAsAction in CI (#34)
* feat(test): add ecdsaSign test, deprecate signAsAction in CI

signAsAction was never implemented (bails immediately server-side) and
has been deprecated per team decision — see Slack discussion linked in
the workflow comment.

Changes:
- k6/lit-action-ecdsa-sign.spec.ts: new test for Lit.Actions.signEcdsa()
  (createAccount → createWallet → signEcdsa with wallet_address as publicKey)
- .github/workflows/k6-lit-action-ecdsa-sign.yml: standalone workflow
- deploy-phala.yml: add k6-lit-action-ecdsa-sign to pipeline; mark
  k6-lit-action-sign-as-action and k6-lit-action-verify-action-signature
  with continue-on-error: true so the deploy pipeline stays green
- k6-lit-action-sign-as-action.yml: add deprecation notice + Slack link
- justfile: add ecdsa-sign to test recipe


* fix(test): get pubkey from listWallets instead of createWallet

createWallet returns only the Ethereum address; signEcdsa's sign_with_pkp
lookup requires the raw uncompressed public key. listWallets returns a
WalletItem with both wallet_address and public_key — use that instead.


* refactor(ci): invert sign-as-action exit code instead of continue-on-error

The workflows now succeed when k6 fails (expected — signAsAction/
verifyActionSignature are not implemented) and fail when k6 passes
(unexpected — would mean signAsAction was accidentally shipped).

This removes continue-on-error from the deploy pipeline entirely;
the reusable workflows handle pass/fail semantics themselves.


* chore: delete verify-action-signature workflow and spec

verifyActionSignature depends on signAsAction which is deprecated and
unimplemented. No need to track it separately.


* chore: remove stray patch file from repo

---------
