+++ .github/workflows/k6-lit-action-sign-as-action.yml
@@ -1,7 +1,19 @@
 # Run lit-action-sign-as-action.spec.ts.
 # Called from Deploy to Phala CVM after wait-for-api, or manually via workflow_dispatch.
 #
-# Required vars: BASE_URL (for workflow_dispatch when base_url not passed)
+# ⚠ EXPECTED TO FAIL — signAsAction is not implemented and has been deprecated.
+#
+# Rationale (from Chris, Slack thread below): anything truly immutable is a bad idea
+# because everyone eventually needs to upgrade, and if they don't want to, they still
+# must because their supply chain (APIs, infra, etc.) changes around them. Everyone
+# creates an "out" anyway, so enforced lock-in adds friction without real benefit.
+# Net/net: easier to move forward without forcibly locking things down.
+#
+# Slack discussion:
+# https://litprotocol.slack.com/archives/C0AG62JLYKS/p1772285399551879?thread_ts=1772254651.978569&cid=C0AG62JLYKS
+#
+# The workflow run succeeds when k6 fails (expected), and fails when k6 passes
+# (unexpected — would mean signAsAction was accidentally implemented).
 
 name: k6 Lit Action Sign As Action
 
@@ -28,7 +40,13 @@ jobs:
       - name: Install k6
         uses: grafana/setup-k6-action@v1
 
-      - name: lit-action-sign-as-action.spec.ts
+      - name: lit-action-sign-as-action.spec.ts (expected to fail)
         env:
           BASE_URL: ${{ inputs.base_url || vars.BASE_URL }}
-        run: k6 run k6/lit-action-sign-as-action.spec.ts
+        run: |
+          if k6 run k6/lit-action-sign-as-action.spec.ts; then
+            echo "UNEXPECTED: signAsAction passed — it should not be implemented"
+            exit 1
+          else
+            echo "OK: signAsAction failed as expected (not implemented, deprecated)"
+          fi
